confirmation modal is replacing the now deleted TeacherDeleteModal, and updated the Teacher settings with this update
